Eliminate `StackBox`, which is effecitvely subsumed by `MoveRef`
